error in php 7

+++ application/models/M_menu.php
@@ -128,7 +128,7 @@ class M_menu extends CI_Model {
           if ($i == 0) {
               // SORTING ASC
                   usort($arr2, function($a, $b) {
-                      return $a['SubMenu1'] - $b['SubMenu1'];
+                      (int)return $a['SubMenu1'] - (int)$b['SubMenu1'];
                   });
           }
